Add lodash, refactor ticket controller
