+++ app/assets/stylesheets/hireo.css
@@ -13,5 +13,4 @@
   *= require users_invitation_email
   *= require comments
   *= require notifications
-  *= require watchers
  */
