+++ app/javascript/bundles/app/ticketShowPage/components/WatchersBlock/index.js
@@ -1,0 +1,3 @@
+import WatchersBlock from './WatchersBlock';
+
+export default WatchersBlock;
