+++ app/javascript/bundles/app/ticketShowPage/components/WatchersBlock/watchers-block.scss
