+++ app/javascript/packs/index.js
@@ -1,6 +1,7 @@
 import ReactOnRails from 'react-on-rails';
 import WatchersBlock from '../bundles/app/ticketShowPage/components/WatchersBlock';
 
+
 ReactOnRails.register({
-  WatchersBlock,
+  WatchersBlock, 
 });
