+++ package.json
@@ -13,6 +13,7 @@
     "bootstrap": "4.4.1",
     "expose-loader": "^0.7.5",
     "jquery": "3.5.0",
+    "lodash": "^4.17.15",
     "popper.js": "^1.16.0",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
